requirements.txt
pydantic==1.8.2
psycopg2-binary==2.9.1
elasticsearch==7.15.1
python-dotenv==0.19.1